Provide a complete sample conductor project
Our first goal will be to perform standard processing of all our
`docker-compose.yml` files.

Note that this project is actually runnable using `docker-compose` if
you use the overlay files:

```sh
docker-compose \
    -f hello/pods/frontend.yml \
    -f hello/pods/overrides/production/frontend.yml \
    up
```

We just need to create a parallel directory structure, and copy
everything over, munging it along the way to fill in the missing
details.