v0.0.6: Official OS X and musl-libc Linux binaries
With this release, we officially provide binaries for OS X and Linux.
The Linux binaries are statically compiled using musl-libc and
rust-musl-builder.  This means that they should work on any reasonably
modern Linux system, including Alpine Linux, which is commonly used in
Docker containers.

Just download, unzip, and copy the binary to wherever you want it.

Alternatively, you can always install using `cargo install conductor`,
as described in the README.