Use handlebars templates for `conductor new`
I'm trying to build a realistic, multi-pod Rails app for our next
example, because we need one to finish specing out the different pod
types.  And I decided that I would use the `new` command to generate
this more complex example.  So I enhanced `new` to support
Handlebars-style templates, because that seems to be the most popular
type of template in the Rust world right now.

CC @dkastner in case he wants to review.