Implement a `rails generate` clone for conductor
This has a nice little re-usable template engine based on
https://github.com/tilpner/includedir.  We do work around one API
limitation described at https://github.com/tilpner/includedir/pull/1.

The interesting details are that we actually compile the entire directory
hierarchy for the generator templates straight into our binary, and that
we can add new files to an existing generator just by putting them in
the right source directory.

This will make it much easier to add new generators quickly, and to add
files to the existing generator.

CC @dkastner