Fix lots of warnings found by "Clippy"
There's a great tool for automatically finding suspicious and
unidiomatic Rust code, and somebody decided to name it "Clippy."  You
can learn more about it here:

https://github.com/Manishearth/rust-clippy

It only had a couple of false positives (which were trivial to work
around), and it taught me some nicer idioms for doing certain things.

To run it, use:

    rustup run nightly cargo install clippy
    rustup run nightly cargo clippy

I think I'm going to turn it on automatically for builds with the
nightly compiler, and convert most warnings into errors.  This will
cause it to fail on Travis CI if we introduce any suspicious code.