Implement a plugin system and a secrets plugin (not yet enabled)
This is a basic "secrets" plugin, for when we just want to prototype
something without setting up a bunch of policies in vault.  Most of the
code here is generic, and we'll recycle it to support other plugins.

This change updates a lot of things.  Some non-obvious bits:

1. We need to enable serde to parse our `config/secrets.yml` file.  This
   requires changes to `build.rs` and some trickery in `Cargo.toml` to
   handle the fact that `docker_compose` also uses serde, and they need
   to agree on whether serde is in macro mode or codegen mode.
2. The plugin system is pretty general.  It's desgined to have several
   kinds of plugins, and the current kind transforms `dc::File` objects
   in arbitrary ways.

Next, we need to hook this plugin up, and make a similar plugin which
handles vault token generation.  This is on the critical path to having
working local development.