Move per-pod configuration into `pods/$NAME.config.yml`
We had some half-backed scheme to treat _consistent_ configuration of
services within a pod as potential configuration for the pod as a
whole.  This was a stupid idea, and it was my fault.  We're just giving
up and creating config files to hold everything we need, since
https://github.com/docker/compose/issues/1655 and
https://github.com/docker/compose/issues/2942 do not show evidence of
happening in the immediate future.