v0.0.8: git branch support, plugins, secrets, vault, lib repos & more
This is a huge update with tons of new features, and I'm releasing it so
that we can test against real projects that are being converted to
conductor.

You will need to replace `io.fdy.conductor.type` fields on your services
with `$POD.config.yml` files containing a `pod_type` field.