Split run and exec args into different structures
These two `docker-compose` subcommands take different argument lists,
and we can't keep trying to jam them into the same structure.

Here, I'm experimenting with `Deref` to provide a primitive sort of
inheritence.  This may go away.

Unfortunately, I can't hook up `-e` yet, because we've reached the
limits of the docopt.rs argument parsing library, and I need to switch
to something better suited to the kind of arguments we want to support.