v0.0.21: Major CLI improvements
Breaking changes in this version:

- All `<POD> <SERVICE>` arguments must now be written as either
  `<POD>/<SERVICE>`, or as just `<SERVICE>` if the name is unique.
- The `cage generate list` command is now called as `cage generate`.

New features include:

- A brand-new argument parser based on clap, with much better subcommand
  support.  Type just `cage` for detailed help.
- Experimental shell completion support for fish and bash.  Run
  `cage generate completion --help` for instructions.
- Support for `--entrypoint` and `-e` arguments.

To migrate an existing project, you will want to add a
`config/project.yml` file containing:

    # This is a semver requirement, and it supports the full syntax at
    # https://github.com/steveklabnik/semver
    cage_version: "0.0.21"

If your version of cage is too old or too new for a given project, it
will print an error.  This will make it easier for teams to keep up with
breaking changes in future releases of cage.