Treat local source code like an always-mounted repo (#24)
This patch is pretty much a textbook example of where Rust wins big.
Before, we had a notion of a `Repo`, which was a git repository
associated with one or more Docker images.  We could clone it locally
and mount it into a Docker image.

But we wanted to support the case where we _didn't_ have an external
repo, and the source code was actually part of our own project.  Now,
`docker-compose` and our `compose_yml` already support this, because the
`build.context` for a service can be either a local directory or a URL.
So there's precedent for what we want.

But we still need to go through our code base, find every piece of code
that manipulates a Git URL, and update it to do the right thing if you
pass it a directory instead.  In a dynamically-typed language, this can
be a long and tedious hunt.

In Rust, all I did was begin by replacing `dc::GitUrl` with
`dc::Context`, and watching for errors.  Every place that caused an
error needed to be redesigned, and the change rippled out from there.
Lots of little APIs needed to be rethought slightly.

Now, dynamic languages do have many advantages.  But static languages
shine when you need to go into a large, pre-existing program and change
a fundamental assumption, because the compiler can ferret out almost
every little corner that needs to be changed.

There will be more patches after this one to add related features and
adjust the CLI.  But this is the core of the refactoring.