Convert ubiquitous Override argument into a Project field (#29)
We want to rename `Override` to `Target`, but unforntunately, it appears
almost _everywhere_ in the API.  I've already noticed that this is a
issue because it makes parameters lists unnecessarily long, and we never
have more than one override active for our standard commands, anyway.

So we move this ubiquitous argument into `Project`, which reduces the
number of APIs that explicitly mention `Override` considerably.