Parallelize pod transforms using rayon (mostly just because)
This should help speed up `cage export` when vault is enabled, too.
There's a good blog post in how Rust repeatedly prevented me from
shooting myself in the foot as I was working on this.

Parallelism has never been so easy!

We have much more important places where parallelism will make a huge
difference in the developer experience, but this was just a warmup
exercise.