Use a more complicated default init script for PG (#34)
So it turns out that we can't just port-scan PostgreSQL and hope that
just because it's listening on port 5432, that it's actually ready to do
business.  PostgreSQL has database recovery modes where it will open the
port and then refuse to accept queries.

We can work around this using a small shell script that calls
`pg_isready`.  This is a hack, but it's very specific to
PostgreSQL--thanks to how systemd handles service initialization these
days, many other services are ready to work as soon as they have an open
port.  So we'll keep the basic design and special-case this.