Bug fix: Make sure `cage test` works (and add test script)
This bug was reported against an internal project, and it represents a
pretty serious regression with the `cage test` command during the
parallel code refactoring.

In order to prevent future regressions of this sort, I'm adding a much
more thorough test script that actually runs the cage binary against
Docker and includes all the standard tutorial steps plus some others.