v0.1.8: Fix builds on older stable Rust
There was an unexpected build failure on Linux for 0.1.7 because it was
using an older version of Rust to make the static release binaries than
it was using to run the regular tests.

This release should restore compatibility with the older Rust we're
using to make binaries (which I also need to update).