v0.1.10: Preliminary dc.yml version 2.1 support & Windows tests pass
We should now have some basic support for many of the features which
will be present in the `docker-compose.yml` format version 2.1, and all
our unit tests are green on Windows.

This does not necessarily mean that Windows works as well as Linux and
Mac, but it should usable for many tasks.  Bug reports and PRs are
eagerly welcome!