cage new: Use Alpine version of PostgreSQL
This is much smaller than the regular PostgreSQL image, and it also
seems to handle `chmod` on mounted data directories better, as far as I
can tell.

Fixes #78, with any luck. See also kiasaki/docker-alpine-postgres#20 for
an example of what we're counting on the standard Docker image to be
able to do correctly.