Cache vault tokens (fixes #100)
Originally, we did not expect people to use the `vault` plugin in
development mode, so we generated new tokens on demand. But this is
time-consuming, and development operations are frequent, so we should
cache them.

We implement a simple cache that renews tokens if half the TTL has
expired or if the desired policies have changed.